Move dependent_services to coreimpl (no code change)
