+++ comp/agents/core/coreimpl/dependent_services.go
@@ -3,7 +3,7 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package run
+package coreimpl
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/util/log"
